Moved StatsCounters model class out of model.py
